refactor: extrai conquistas e estatísticas do TrailController para services
Move o catálogo de conquistas (CRUD), a premiação automática, o catálogo marcado do usuário e o feed da comunidade para achievement.service. As estatísticas de XP, aulas e acertos vão para stats.service, reusado pelo getMyXp e pela premiação. Mesmo comportamento da API.
