+++ backend/src/services/stats.service.ts
@@ -1,0 +1,22 @@
+import { db } from "../../db.ts";
+import { lessonProgress, questionAnswers } from "../../schema.ts";
+import { eq, and, count } from "drizzle-orm";
+
+// Estatísticas base do usuário. XP = 50 por aula concluída + 10 por questão certa.
+export async function calcularEstatisticas(userId: string) {
+    const [aulas] = await db
+        .select({ n: count() })
+        .from(lessonProgress)
+        .where(eq(lessonProgress.userId, userId));
+    const [acertos] = await db
+        .select({ n: count() })
+        .from(questionAnswers)
+        .where(and(eq(questionAnswers.userId, userId), eq(questionAnswers.isCorrect, true)));
+    const lessonsCompleted = Number(aulas?.n ?? 0);
+    const questionsCorrect = Number(acertos?.n ?? 0);
+    return {
+        xp: lessonsCompleted * 50 + questionsCorrect * 10,
+        lessonsCompleted,
+        questionsCorrect,
+    };
+}
